+++ apps/valence-app/src/app/programs/ui/CustomChainForm.tsx
@@ -15,7 +15,7 @@ import { ChainInfo } from "@keplr-wallet/types";
 import { useForm } from "react-hook-form";
 import { neutron as neutronChainInfo } from "graz/chains";
 import { checkWallet, WalletType } from "graz";
-import { supportedChains } from "@/context";
+import { programsSupportedChains } from "@/const";
 
 export type CustomChainFormValues = {
   suggestedChainInfo: string;
@@ -33,7 +33,7 @@ export const CustomChainForm = ({
   onSubmit: (chainInfo: ChainInfo) => void;
 }) => {
   const suggestedChainInfo =
-    supportedChains.find((chain) => chain.chainName === domainName) ??
+    programsSupportedChains.find((chain) => chain.chainName === domainName) ??
     neutronChainInfo;
   const {
     currencies,
